# lib/api-zod/src/generated/types/attendance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceMethod } from './attendanceMethod';
import type { AttendanceStatus } from './attendanceStatus';

export interface Attendance {
  id: number;
  memberId: number;
  memberName: string;
  memberCode: string;
  date: Date;
  time: string;
  status: AttendanceStatus;
  method: AttendanceMethod;
}