# lib/api-zod/src/generated/types/checkInInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CheckInInputMethod } from './checkInInputMethod';

export interface CheckInInput {
  memberId: number;
  method?: CheckInInputMethod;
}