# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { DashboardGym } from './dashboardGym';
import type { DashboardMetrics } from './dashboardMetrics';
import type { SeriesPoint } from './seriesPoint';

export interface Dashboard {
  gym: DashboardGym;
  metrics: DashboardMetrics;
  revenue: SeriesPoint[];
  attendance: SeriesPoint[];
  membershipMix: SeriesPoint[];
  recentActivity: Activity[];
}