# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  name: string;
  category: string;
  amount: number;
  date: Date;
  method: string;
  /** @nullable */
  description?: string | null;
}