# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  name: string;
  category: string;
  /** @minimum 0 */
  amount: number;
  method: string;
  /** @nullable */
  description?: string | null;
}