# lib/api-zod/src/generated/types/memberUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MemberInput } from './memberInput';

export type MemberUpdate = MemberInput;