# lib/api-zod/src/generated/types/membershipPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { MembershipPlanStatus } from './membershipPlanStatus';

export interface MembershipPlan {
  id: number;
  name: string;
  duration: string;
  price: number;
  memberCount: number;
  status: MembershipPlanStatus;
  color: string;
}