# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethod } from './paymentMethod';
import type { PaymentStatus } from './paymentStatus';

export interface Payment {
  id: number;
  receiptNumber: string;
  memberName: string;
  memberCode: string;
  amount: number;
  method: PaymentMethod;
  status: PaymentStatus;
  date: Date;
  planName: string;
  discount?: number;
}