# lib/api-zod/src/generated/types/trainer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Gym Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TrainerStatus } from './trainerStatus';

export interface Trainer {
  id: number;
  name: string;
  specialization: string;
  memberCount: number;
  status: TrainerStatus;
  initials: string;
  phone: string;
  email: string;
  schedule?: string;
}